Adds additional candle fields.
Additional fields are:
   . buyVolume
   . buyTrades
   - lag (exchange lag)

These new fields are added to candle data only if `config.candleWriter.version`
field is set to 2. If it is 1 or if it is not set, then candles will be created
with the same values as before.

Note that you need to add these fields to database if you have
table already created. New tables get these fields automatically, if
candle version 2 is set.

The patch makes it easy if we ever need to add another set of
new fields. All that is needed is to increment version number and
add all the needed if()s with new field values.